data_format_fixes

+++ source/functions/BEMIO/normalizeBEM.m
@@ -56,30 +56,30 @@ if issorted(hydro(F).w)==0  % Sort, if necessary
 end
 
 if strcmp(hydro(F).code,'WAMIT')==0  % normalize
-    hydro(F).Khs = hydro(F).Khs/(hydro(F).g*hydro(F).rho);
-    hydro(F).A = hydro(F).A/(hydro(F).rho);
+    hydro(F).Khs = hydro(F).Khs/double(hydro(F).g*hydro(F).rho);
+    hydro(F).A = hydro(F).A/double(hydro(F).rho);
     hydro(F).Ainf = hydro(F).A(:,:,end); % overwritten with more accurate method by radiationIRF.m 
     for i=1:length(hydro(F).w)
-        hydro(F).B(:,:,i) = hydro(F).B(:,:,i)/(hydro(F).rho*hydro(F).w(i));
+        hydro(F).B(:,:,i) = hydro(F).B(:,:,i)/double(hydro(F).rho*hydro(F).w(i));
     end
-    hydro(F).ex_ma = hydro(F).ex_ma/(hydro(F).g*hydro(F).rho);
-    hydro(F).ex_re = hydro(F).ex_re/(hydro(F).g*hydro(F).rho);
-    hydro(F).ex_im = hydro(F).ex_im/(hydro(F).g*hydro(F).rho);
+    hydro(F).ex_ma = hydro(F).ex_ma/double(hydro(F).g*hydro(F).rho);
+    hydro(F).ex_re = hydro(F).ex_re/double(hydro(F).g*hydro(F).rho);
+    hydro(F).ex_im = hydro(F).ex_im/double(hydro(F).g*hydro(F).rho);
     if isfield(hydro(F),'md_mc') 
-        hydro(F).md_mc = hydro(F).md_mc/(hydro(F).g*hydro(F).rho);
+        hydro(F).md_mc = hydro(F).md_mc/double(hydro(F).g*hydro(F).rho);
     end
     if isfield(hydro(F),'md_cs')    
-        hydro(F).md_cs = hydro(F).md_cs/(hydro(F).g*hydro(F).rho);    
+        hydro(F).md_cs = hydro(F).md_cs/double(hydro(F).g*hydro(F).rho);    
     end
     if isfield(hydro(F),'md_pi')
-        hydro(F).md_pi = hydro(F).md_pi/(hydro(F).g*hydro(F).rho);
+        hydro(F).md_pi = hydro(F).md_pi/double(hydro(F).g*hydro(F).rho);
     end        
-    hydro(F).sc_ma = hydro(F).sc_ma/(hydro(F).g*hydro(F).rho);
-    hydro(F).sc_re = hydro(F).sc_re/(hydro(F).g*hydro(F).rho);
-    hydro(F).sc_im = hydro(F).sc_im/(hydro(F).g*hydro(F).rho);
-    hydro(F).fk_ma = hydro(F).fk_ma/(hydro(F).g*hydro(F).rho);
-    hydro(F).fk_re = hydro(F).fk_re/(hydro(F).g*hydro(F).rho);
-    hydro(F).fk_im = hydro(F).fk_im/(hydro(F).g*hydro(F).rho);
+    hydro(F).sc_ma = hydro(F).sc_ma/double(hydro(F).g*hydro(F).rho);
+    hydro(F).sc_re = hydro(F).sc_re/double(hydro(F).g*hydro(F).rho);
+    hydro(F).sc_im = hydro(F).sc_im/double(hydro(F).g*hydro(F).rho);
+    hydro(F).fk_ma = hydro(F).fk_ma/double(hydro(F).g*hydro(F).rho);
+    hydro(F).fk_re = hydro(F).fk_re/double(hydro(F).g*hydro(F).rho);
+    hydro(F).fk_im = hydro(F).fk_im/double(hydro(F).g*hydro(F).rho);
 end
 
 end
